Update patch set 2
Patch Set 2:

FWIW, the release/3.0 branch has seen 4 releases with a large number (325) of changes since this was created.  Those are mostly bug fixes.

The release branch merges cleanly into our code so I am going to build the latest release/3.0 (which is version 3.0.8) and compare the CTS results against what we have now.  If it is an improvement I'll push the squash up for others to test.

Patch-set: 2