Update patch set 2
Patch Set 2:

I the CTS tests I mentioned on pme both before and after updating the code.  The failures were the same which means that updating ffmpeg neither fixes any issues detected by cts nor regresses any.

The diffs in our version of ffmpeg vs. upstream are small (aside from Android makefiles/configs).

Therefore I think we should update this to keep current with upstream of get all the stable fixes pushed to their 3.0 release.

I'm not going to bother to update this commit with the latest code, it is at:

https://github.com/crpalmer/android_external_ffmpeg/commits/cm-14.1-release-3.0.8

if anyone wants to test it or push it to our lineage branch.

Patch-set: 2